bunch of changes
We test to see if another instance is running (useful for cronjobs that take a while to complete)
We test to make sure we can enter the target directory.
We use a facecreep.rc.txt file to track variables (EXCEPT directory, which points to that file).
Format of facecreep.rc.txt is newline delimited email, password, target account, and your username as it appears in facebook (for testing if you're logged in)
This regex is bullshit, I keep having to shift through the photos to get the right one. For instance they're all set to 3 now and that works for me.